minor fixes and modularized
